# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFtaXJleDEyOEBnbWFpbC5jb20iLCJleHAiOjUyNjgzMTg5OTYsImV4cGlyZV9hdCI6IiIsImZpcnN0bmFtZSI6Itin2YXbjNixIiwiaWQiOjEsImxhc3RuYW1lIjoi2LTbjNix2K_ZhNuMIiwibW9iaWxlIjoiMDkwMjQ4MDk3NTAiLCJvcmlnX2lhdCI6MTY2ODMyMjU5Niwic3RhdHVzIjoiIn0.x7BKuxw288cm1JsskGRD178UPmNz-xRwkWHtb0WsU74",
        "title": "Selloora Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://www.amirshirdel.ir",
            "email": "amirex128@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/",
    "paths": {
...
        "/user/shop": {
            "get": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "لیست فروشگاه",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "متن جستجو",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "شماره صفحه",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "تعداد صفحه",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "مرتب سازی براساس desc/asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/user/shop/create": {
            "post": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "ایجاد فروشگاه",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateShop"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/shop/delete/{id}": {
            "post": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "حذف فروشگاه",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه فروشگاه",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/shop/send-price": {
            "post": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "ویرایش هزینه ارسال سفارشات به صورت جدا",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.SendPrice"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/shop/show/{id}": {
            "get": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "نمایش فروشگاه",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه فروشگاه",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/shop/update": {
            "post": {
                "description": "هر کاربر برای این که بتواند محصولی ایجاد کند باید فروشگاه داشته باشد تا محصولات و مقالات خود را بر روی این فروشگاه ذخیره کند این فروشگاه میتواند ربات تلگرام باشد یا سایت باشد یک نمونه مشابه اینستاگرام باشد",
                "tags": [
                    "shop"
                ],
                "summary": "ویرایش فروشگاه",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.UpdateShop"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "DTOs.CreateShop": {
            "type": "object",
            "required": [
                "name",
                "social_address",
                "type"
            ],
            "properties": {
                "description": {
                    "description": "توضیحات فروشگاه",
                    "type": "string",
                    "example": "توضیحات فروشگاه"
                },
                "email": {
                    "description": "ایمیل برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128@gmail.com"
                },
                "gallery_id": {
                    "description": "تصویر لوگو فروشگاه",
                    "type": "integer",
                    "example": 1
                },
                "instagram_id": {
                    "description": "آیدی اینستاگرام برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex_dev"
                },
                "mobile": {
                    "description": "شماره موبایل برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "09024809750"
                },
                "name": {
                    "description": "نام فروشگاه",
                    "type": "string",
                    "example": "فروشگاه امیر"
                },
                "phone": {
                    "description": "شماره تلفن برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "05136643278"
                },
                "send_price": {
                    "description": "هزینه ارسال تومان",
                    "type": "number",
                    "example": 20000
                },
                "social_address": {
                    "description": "آدرس شبکه اجتماعیی که میخواهید متصل کنید به فروشگاه برای دریافت محصولات از ان",
                    "type": "string",
                    "example": "amirex_dev"
                },
                "telegram_id": {
                    "description": "آیدی تلگرام برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128"
                },
                "theme_id": {
                    "description": "شناسه قالب",
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "description": "نوع فروشگاه که میتواند به صورت اینستاگرامی باشد یا وبسایتی باشد یا روبیکا یا ربات تلگرام",
                    "type": "string",
                    "enum": [
                        "instagram",
                        "telegram",
                        "website",
                        "rubika"
                    ],
                    "example": "instagram"
                },
                "website": {
                    "description": "وبسایت شخصی برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "https://amirshirdel.ir"
                },
                "whatsapp_id": {
                    "description": "آیدی واتس آپ برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128"
                }
            }
        },
...
        "DTOs.SendPrice": {
            "type": "object",
            "required": [
                "send_price",
                "shop_id"
            ],
            "properties": {
                "send_price": {
                    "description": "هزینه ارسال تومان",
                    "type": "number",
                    "example": 30000
                },
                "shop_id": {
                    "description": "شناسه فروشگاه",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "DTOs.UpdateShop": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "description": "توضیحات فروشگاه",
                    "type": "string",
                    "example": "توضیحات فروشگاه"
                },
                "email": {
                    "description": "ایمیل برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128@gmail.com"
                },
                "gallery_id": {
                    "description": "تصویر لوگو فروشگاه",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "شناسه فروشگاه برای ویرایش",
                    "type": "integer",
                    "example": 1
                },
                "instagram_id": {
                    "description": "آیدی اینستاگرام برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex_dev"
                },
                "mobile": {
                    "description": "شماره موبایل برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "09024809750"
                },
                "name": {
                    "description": "نام فروشگاه",
                    "type": "string",
                    "example": "فروشگاه امیر"
                },
                "phone": {
                    "description": "شماره تلفن برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "05136643278"
                },
                "send_price": {
                    "description": "هزینه ارسال تومان",
                    "type": "number",
                    "example": 25000
                },
                "social_address": {
                    "description": "آدرس شبکه اجتماعیی که میخواهید متصل کنید به فروشگاه برای دریافت محصولات از ان",
                    "type": "string",
                    "example": "amirex_dev"
                },
                "telegram_id": {
                    "description": "آیدی تلگرام برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128"
                },
                "theme_id": {
                    "description": "شناسه قالب",
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "description": "نوع فروشگاه که میتواند به صورت اینستاگرامی باشد یا وبسایتی باشد یا روبیکا یا ربات تلگرام",
                    "type": "string",
                    "enum": [
                        "instagram",
                        "telegram",
                        "website",
                        "rubika"
                    ],
                    "example": "instagram"
                },
                "website": {
                    "description": "وبسایت شخصی برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "https://amirshirdel.ir"
                },
                "whatsapp_id": {
                    "description": "آیدی واتس آپ برای نمایش در فروشگاه برای پشتیبانی",
                    "type": "string",
                    "example": "amirex128"
                }
            }
        },
...
    }
}